Code review feedback and fixes
